server/requirements.txt
django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-environ==0.11.2
python-dotenv==1.0.1
mysqlclient==2.2.7
Pillow==12.0.0
tensorflow==2.13.0
opencv-python==4.10.1.26
scikit-learn==1.5.0
numpy==1.26.0
pandas==2.2.0
scipy==1.13.0
requests==2.31.0
